# contracts/openapi/ztax.v1.yaml
...
openapi: 3.2.0
info:
  title: ZoikoTax
  version: "1.0.0"
  summary: Tenant identity, session and administration surface of a ZoikoTax regional cell.
  description: |
    This is the `/v1` surface a ZoikoTax regional cell serves today: discovery,
    authentication and tenant administration. The determination surface named in
    `ZTAX-INT-001` — quotes, transaction commit, adjust, refund, decisions,
    replay, classifications, obligations, batches and jobs — is not here yet and
    is added by addition under the same major version (ADR-0010 §2.6).

    **No authoritative fiscal output is permitted before A4.** A client
    discovers whether this deployment may produce one from
    `GET /v1/capabilities`, which reports `authoritative`, never from a release
    note. A deployment reporting `authoritative: false` produces advisory
    figures that must not be filed.

    ## Errors

    Every error is an RFC 9457 Problem Details document with
    `Content-Type: application/problem+json` (ADR-0016). Two fields matter more
    than the status code:

    - `ztx_reason_code` is from a closed, registered vocabulary. It means the
      same thing in an error, in a decision and in evidence, and it is what a
      client branches on. Never match on `title` or `detail`; both may be
      reworded.
    - `ztx_retryable` states whether retrying the request unchanged could
      succeed. It is a property of the error rather than something to infer from
...

...
    `GET /v1/capabilities`.

...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
  /v1/capabilities:
    get:
      operationId: getCapabilities
      tags: [discovery]
      summary: Report this deployment's effective capability
      description: |
        Effective capability, not entitlement: the intersection of content
        state, security, QA and commercial entitlement (ADR-0010 §2.6).

        Unauthenticated, because a client needs to know whether a cell can serve
        it before it has a session, and because an integration diagnosing a
        problem should not need a credential to report which cell and which
        content bundle it reached.

        `content` is absent in a cell with no bundle loaded, which is a
        legitimate deployment before A4. Absent rather than null: "no content"
        and "content with no identity" are different facts (ADR-0011 P3).
      security: []
      responses:
        "200":
          description: The deployment's effective capability.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
              examples:
                preAuthorization:
                  summary: A cell before A4, with a worked pack loaded
                  value:
                    cell: eu-west-1
                    region: eu-west
                    environment: development
                    trains:
                      app: 0.4.0
                      content: 0.0.0-dev
                      ai: 0.1.0
                      adapter: 0.2.0
                      infra: 0.3.0
                      schema: 0.1.0
                      migration: "6"
                    canonProfile: canon/v1
                    authoritative: false
                    reasonCodes: [AMBIGUOUS, CURRENCY_MISMATCH, FORBIDDEN, NOT_AUTHORITATIVE, NO_CONTENT_BUNDLE]
                    content:
                      bundleId: eu-vat-worked-2026.09
                      digest: "zt1:be8651b55057f0f11f2a39820495b373417e5aa81b122f61b9fcb79faaf86c7b"
                      irVersion: 1
                      nodeCount: 22
                noContent:
                  summary: A cell deployed with no content bundle
                  value:
                    cell: eu-west-1
                    region: eu-west
                    environment: production
                    trains:
                      app: 0.4.0
                      content: none
                      ai: 0.1.0
                      adapter: 0.2.0
                      infra: 0.3.0
                      schema: 0.1.0
                      migration: "6"
                    canonProfile: canon/v1
                    authoritative: false
                    reasonCodes: [NO_CONTENT_BUNDLE, NOT_AUTHORITATIVE]
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/auth/sign-in:
    post:
      operationId: signIn
      tags: [authentication]
      summary: Open a session
      description: |
        Exchanges a tenant slug, an email address and a password for an opaque
        session cookie.

        The response carries **no token**. The session identifier is in a
        `__Host-`-prefixed, `HttpOnly`, `SameSite=Strict` cookie that JavaScript
        cannot read, which is why the browser client has nothing to store and
        nothing to leak (ADR-0019 C8, ADR-0020).

        A failure does not distinguish an unknown subject from a wrong password.
        The difference is an account-enumeration oracle and a caller can do
        nothing with it, so both are `INVALID_CREDENTIALS`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
            examples:
              default:
                summary: An administrator signing in
                value:
                  tenant: acme
                  email: admin@acme.example
                  password: "correct horse battery staple"
      responses:
        "200":
          description: The session, and the subject it belongs to.
          headers:
            Set-Cookie:
              description: |
                The opaque session cookie. `__Host-ztax_session` where the
                deployment is served over TLS; `ztax_session` on plain HTTP in
                local development only, because a browser silently drops a
                `__Host-` cookie that is not `Secure`.
              schema: { type: string }
              example: "__Host-ztax_session=…; Path=/; Secure; HttpOnly; SameSite=Strict"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
              examples:
                default:
                  summary: A signed-in administrator
                  value:
                    tenant:
                      id: ztn_01JBQ0S9C3X8Q1H6M2KX5R7F4A
                      slug: acme
                      displayName: Acme Telecom
                      residencyRegion: eu-west
                      status: ACTIVE
                    user:
                      id: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4B
                      email: admin@acme.example
                      displayName: Ada Lovelace
                      status: ACTIVE
                      roles: [ADMIN]
                      createdAt: "2026-09-01T09:14:22.104000Z"
                    expiresAt: "2026-09-24T09:14:22.104000Z"
        "400": { $ref: "#/components/responses/Validation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/auth/sign-out:
    post:
      operationId: signOut
      tags: [authentication]
      summary: End the current session
      description: |
        Revokes this session server-side and clears the cookie. The session is
        revoked rather than forgotten, so it appears in
        `GET /v1/admin/sessions` with `revoked: true` — a session that vanished
        would leave nothing for an audit to look at.
      responses:
        "204":
          description: The session is revoked and the cookie is cleared.
          headers:
            Set-Cookie:
              description: The cleared session cookie.
              schema: { type: string }
              example: "__Host-ztax_session=; Path=/; Max-Age=0; Secure; HttpOnly; SameSite=Strict"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/auth/session:
    get:
      operationId: getSession
      tags: [authentication]
      summary: Who am I, and until when
      description: |
        The only way a client learns its own identity. Nothing about the session
        is readable in the browser, so a client calls this on load rather than
        decoding a token (ADR-0019 C8).
      responses:
        "200":
          description: The current session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
              examples:
                default:
                  summary: An analyst's session
                  value:
                    tenant:
                      id: ztn_01JBQ0S9C3X8Q1H6M2KX5R7F4A
                      slug: acme
                      displayName: Acme Telecom
                      residencyRegion: eu-west
                      status: ACTIVE
                    user:
                      id: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4C
                      email: analyst@acme.example
                      displayName: Grace Hopper
                      status: ACTIVE
                      roles: [ANALYST]
                      createdAt: "2026-09-02T11:02:00.000000Z"
                    expiresAt: "2026-09-24T09:14:22.104000Z"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/auth/password:
    post:
      operationId: changePassword
      tags: [authentication]
      summary: Change the current subject's password
      description: |
        Verifies the current password and replaces it.

        **Every session for this subject is revoked, including this one.** A
        password change is the action taken when a credential is believed
        compromised, and leaving other sessions alive would defeat it. The
        client is signed out and must sign in again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
            examples:
              default:
                summary: Rotating a password
                value:
                  currentPassword: "correct horse battery staple"
                  newPassword: "a considerably longer passphrase than that one"
      responses:
        "204":
          description: The password is changed and every session is revoked.
        "400": { $ref: "#/components/responses/Validation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/tenant:
    get:
      operationId: getTenant
      tags: [administration]
      summary: The authenticated subject's tenant
      description: |
        Requires a session but no particular role: every member of a tenant may
        see which tenant they are in, its residency region and whether it is
        suspended.
      responses:
        "200":
          description: The tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
              examples:
                default:
                  summary: An active tenant
                  value:
                    id: ztn_01JBQ0S9C3X8Q1H6M2KX5R7F4A
                    slug: acme
                    displayName: Acme Telecom
                    residencyRegion: eu-west
                    status: ACTIVE
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/users:
    get:
      operationId: listUsers
      tags: [administration]
      summary: List the tenant's users
      description: Readable by `ADMIN`, `ANALYST` and `AUDITOR`.
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The tenant's users.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
              examples:
                default:
                  summary: Two users
                  value:
                    users:
                      - id: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4B
                        email: admin@acme.example
                        displayName: Ada Lovelace
                        status: ACTIVE
                        roles: [ADMIN]
                        createdAt: "2026-09-01T09:14:22.104000Z"
                      - id: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4C
                        email: analyst@acme.example
                        displayName: Grace Hopper
                        status: INVITED
                        roles: [ANALYST]
                        createdAt: "2026-09-02T11:02:00.000000Z"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }

    post:
      operationId: createUser
      tags: [administration]
      summary: Create a user
      description: |
        `ADMIN` only. A user created without a password is `INVITED` and cannot
        sign in until one is set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              invited:
                summary: An invited auditor with no password yet
                value:
                  email: auditor@acme.example
                  displayName: Katherine Johnson
                  roles: [AUDITOR]
      responses:
        "201":
          description: The created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                default:
                  summary: The invited auditor
                  value:
                    id: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4D
                    email: auditor@acme.example
                    displayName: Katherine Johnson
                    status: INVITED
                    roles: [AUDITOR]
                    createdAt: "2026-09-23T12:00:00.000000Z"
        "400": { $ref: "#/components/responses/Validation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/users/{userId}/status:
    post:
      operationId: setUserStatus
      tags: [administration]
      summary: Enable or disable a user
      description: |
        `ADMIN` only. Disabling a user revokes their sessions: an account whose
        access has been withdrawn and whose session still works is an account
        whose access has not been withdrawn.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserStatusRequest"
            examples:
              disable:
                summary: Withdrawing access
                value:
                  status: DISABLED
      responses:
        "204":
          description: The status is set.
        "400": { $ref: "#/components/responses/Validation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/users/{userId}/roles:
    post:
      operationId: grantRole
      tags: [administration]
      summary: Grant a role
      description: "`ADMIN` only. Granting a role a user already holds is not an error."
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleRequest"
            examples:
              default:
                summary: Granting the auditor role
                value:
                  role: AUDITOR
      responses:
        "204":
          description: The role is granted.
        "400": { $ref: "#/components/responses/Validation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/users/{userId}/roles/{role}:
    delete:
      operationId: revokeRole
      tags: [administration]
      summary: Revoke a role
      description: |
        `ADMIN` only. Revoking the last `ADMIN` role in a tenant is refused with
        `STATE_TRANSITION_INVALID`: a tenant with no administrator cannot be
        administered, and recovering from it is a support incident.
      parameters:
        - $ref: "#/components/parameters/UserId"
        - name: role
          in: path
          required: true
          description: The role to revoke.
          schema: { $ref: "#/components/schemas/Role" }
          example: AUDITOR
      responses:
        "204":
          description: The role is revoked.
        "400": { $ref: "#/components/responses/Validation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/sessions:
    get:
      operationId: listSessions
      tags: [administration]
      summary: List the tenant's sessions
      description: |
        `ADMIN` only. `current` marks the caller's own session, so an
        administrative screen can warn before revoking it rather than signing
        the administrator out with no explanation.
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The tenant's sessions.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/SessionSummary" }
              examples:
                default:
                  summary: The caller's session and a revoked one
                  value:
                    sessions:
                      - id: zts_01JBQ0S9C3X8Q1H6M2KX5R7F4E
                        userId: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4B
                        createdAt: "2026-09-23T09:14:22.104000Z"
                        expiresAt: "2026-09-24T09:14:22.104000Z"
                        userAgent: Mozilla/5.0
                        clientIp: 203.0.113.7
                        revoked: false
                        current: true
                      - id: zts_01JBQ0S9C3X8Q1H6M2KX5R7F4F
                        userId: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4C
                        createdAt: "2026-09-20T08:00:00.000000Z"
                        expiresAt: "2026-09-21T08:00:00.000000Z"
                        revoked: true
                        current: false
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/sessions/{sessionId}:
    delete:
      operationId: revokeSession
      tags: [administration]
      summary: Revoke a session
      description: "`ADMIN` only. The session is marked revoked, never deleted."
      parameters:
        - name: sessionId
          in: path
          required: true
          description: The session identifier.
          schema: { $ref: "#/components/schemas/SessionId" }
          example: zts_01JBQ0S9C3X8Q1H6M2KX5R7F4F
      responses:
        "204":
          description: The session is revoked.
        "400": { $ref: "#/components/responses/Validation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/admin/audit:
    get:
      operationId: listAudit
      tags: [administration]
      summary: Read the tenant's audit trail
      description: |
        Readable by `ADMIN` and `AUDITOR`. Append-only: records are written, never
        amended, and a correction is a new record (ADR-0003).

        `detail` is a canonical JSON document as a string. It is a string rather
        than an object because it is the exact bytes that were recorded and
        digested — re-encoding it through this API would mean the value a reader
        sees is this build's serializer's opinion of what was written, rather
        than what was written (ADR-0011 §2.7).
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The audit records, most recent first.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [records]
                properties:
                  records:
                    type: array
                    items: { $ref: "#/components/schemas/AuditRecord" }
              examples:
                default:
                  summary: A role grant
                  value:
                    records:
                      - id: zta_01JBQ0S9C3X8Q1H6M2KX5R7F4G
                        actorUserId: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4B
                        action: user.role.granted
                        subjectType: user
                        subjectId: ztu_01JBQ0S9C3X8Q1H6M2KX5R7F4C
                        detail: '{"role":"AUDITOR"}'
                        recordedAt: "2026-09-23T12:00:00.000000Z"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ztax_session
      description: |
        An opaque, server-side session. The cookie is `HttpOnly` and
        `SameSite=Strict`, so it is sent by a browser automatically and is not
        readable by script. There is no bearer token and no JWT: a token a
        client can read is a token a client can leak, and a stateless token
        cannot be revoked (ADR-0020).

        On plain HTTP in local development the cookie is named `ztax_session`,
        because a browser silently drops a `__Host-` cookie that is not
        `Secure`.
...
  responses:
...
    Internal:
      description: |
        A defect. The cause is recorded and alerted; it is deliberately not in
        the response, because an error string from a driver carries table names,
        query fragments and values (ADR-0016 §2.6).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            internal:
              summary: An unclassified failure
              value:
                type: https://errors.zoikotax.com/v1/internal-error
                title: Internal error
                status: 500
                detail: An unexpected condition prevented the request from completing.
                instance: /v1/admin/users
                ztx_reason_code: INTERNAL_ERROR
                ztx_request_id: 01JBQ0S9C3X8Q1H6M2KX5R7F4P
                ztx_retryable: false
...